refactor(frontend): unify Rule types, extract RuleEditor component
Create base Rule interface extended by SwitchRule (with label). Extract
shared rule editing UI from NodeDetailsPanel into reusable RuleEditor
component used by both switch and filter sections. Add assertion to
ComponentType and NodePalette ICONS for upcoming assertion node.
